SQL-retrieval-Test (Gen-AI)/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
chromadb
sqlite